refactor: test API를 위한 별도 DummyController 생성 (추후 삭제 예정)
