+++ back-end/src/main/java/com/tdd/backend/user/UserController.java
@@ -12,13 +12,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.tdd.backend.auth.LoginUser;
 import com.tdd.backend.user.data.UserCreate;
 import com.tdd.backend.user.data.UserLogin;
-import com.tdd.backend.user.data.UserSession;
 import com.tdd.backend.user.exception.DuplicateEmailException;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -67,20 +64,5 @@ public class UserController {
 			.location(URI.create("/"))
 			.header(HttpHeaders.SET_COOKIE, cookie.toString())
 			.build();
-	}
-
-	@Operation(summary = "CORS 테스트를 위한 더미 API", description = "인증 필요없이 접근 가능")
-	@RequestMapping("/test")
-	public String test() {
-		return "hello world";
-	}
-
-	@Operation(summary = "CORS 테스트를 위한 더미 API (인증필요)", description = "인증이 요구되는 접근 (쿠키 필요)")
-	@RequestMapping("/test/auth")
-	public void testAuth(@LoginUser UserSession session) {
-
-		User user = session.getUser();
-		log.info(">> user : {}", user.getUserName());
-		log.info(">> token : {}", session.getAccessToken());
 	}
 }
